requirements.txt
langchain_openai 
langchain_core
python-dotenv
streamlit
langchain_community
langchain_pinecone
langserve
faiss-cpu
groq
beautifulsoup4
langchain-groq
wikipedia
arxiv
langchainhub
sentence_transformers
PyPDF2
pypdf
pdfplumber
flask
pinecone-client